Pr/2

+++ libudis86/decode.c
@@ -669,7 +669,7 @@ decode_modrm_rm(struct ud         *u,
   }
 
   if (offset) {
-    decode_mem_disp(u, offset, op);
+    decode_mem_disp(u, (unsigned int) offset, op);
   } else {
     op->offset = 0;
   }
@@ -1278,7 +1278,7 @@ ud_decode(struct ud *u)
   u->pc += u->inp_ctr;    /* move program counter by bytes decoded */
 
   /* return number of bytes disassembled. */
-  return u->inp_ctr;
+  return (unsigned int) u->inp_ctr;
 }
 
 /*
